pipeline/requirements.txt
absl-py==2.1.0
accelerate==0.30.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
attrs==23.2.0
braceexpand==0.1.7
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.23.0
humanize==4.9.0
idna==3.4
ignite==1.1.0
imageio==2.34.1
Jinja2==3.1.3
joblib==1.4.2
kaleido==0.2.1
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
MedPy==0.5.1
ml-dtypes==0.3.2
monai==1.3.0
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
networkx==3.2.1
nibabel==5.2.1
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pathtools==0.1.2
pillow==10.2.0
platformdirs==4.2.1
plotly==5.22.0
pretty-errors==1.2.25
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyDeprecate==0.3.2
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==1.7.7
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.10
requests==2.28.1
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
sentry-sdk==2.1.1
setproctitle==1.3.3
sharedmem==0.3.8
shellingham==1.5.4
silence_tensorflow==1.2.1
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
support_developer==1.0.5
sympy==1.12
tabulate==0.9.0
taming-transformers==0.0.1
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-gan==2.1.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-probability==0.24.0
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tifffile==2024.5.3
timm==0.9.16
tokenizers==0.19.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchio==0.19.6
torchmetrics==0.11.4
torchvision==0.16.0+cu121
tqdm==4.66.4
transformers==4.40.2
triton==2.1.0
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
wandb==0.17.0
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
torch-fidelity==0.3.0
